cell_sort 向量化(4.7s/窗口, 数组位级=参考循环) + 闭式化尝试被测试抓住回退的记录
- 布尔矩阵按 C 列并集精确去重(递增序=排序) + int32 三候选重叠选择
  (首最大平局) + ragged 笛卡尔积; 枚举序与参考循环逐位一致
- CI: 向量化 vs 参考数组位级等价 (seed x tc) + 既有放置/owner 测试
- 闭式化 (去 s=-1 halo 候选 + s=+1 单判据) 被 placement[12-2]/owner
  测试当场抓住回退——seam 角 flush 相关 staging 被改变; 失败模式登记
- pod 复跑: forall-K 10.2s 通过; e0n/e0o 全绿

+++ gpu/pme.py
@@ -468,9 +468,14 @@ def _cell_sort_vec(x: np.ndarray, box, ng: int, tc: int):
         B[d][ai, rank[ai, bi] - 1] = bi       # compact, increasing order
 
     # per-dim vectorized shift selection over the padded blocks (int32
-    # domain; identical first-max tie-break as cell_sort_reference)
+    # domain; identical first-max tie-break as cell_sort_reference).
+    # NOTE a closed-form replacement (drop s=-1 as "halo-only", select
+    # s=+1 by its overlap>0 alone) was REVERTED: it changed flush-relevant
+    # staging in seam corners (placement/owner tests caught it) -- the
+    # three-candidate max-overlap with the exact first-max tie-break is
+    # the verified form.  Further speedup must preserve it exactly.
     def shift_for(d: int) -> np.ndarray:
-        ka = au[:, :, d]                                  # (n, R)
+        ka = au[:, :, d].astype(np.int32)                 # (n, R)
         ox = (B[d] * tc).astype(np.int32)                 # (n, C)
         ka32 = ka.astype(np.int32)                        # (n, R)
         lo0 = (ka32[:, None, :] - 3)                      # (n, 1, R)
